Thank god

+++ tenmo-server/src/main/java/com/techelevator/tenmo/model/Account.java
@@ -4,23 +4,23 @@ import java.math.BigDecimal;
 
 public class Account {
 
-    private Long account_id;
-    private Long user_id;
+    private int account_id;
+    private int user_id;
     private BigDecimal balance;
 
     public Account() {}
 
-    public Account(Long account_id, Long user_id, BigDecimal balance) {
+    public Account(int account_id, int user_id, BigDecimal balance) {
         this.account_id = account_id;
         this.user_id = user_id;
         this.balance = balance;
     }
 
-    public Long getAccount_id() {
+    public int getAccount_id() {
         return account_id;
     }
 
-    public Long getUser_id() {
+    public int getUser_id() {
         return user_id;
     }
 
@@ -28,11 +28,11 @@ public class Account {
         return balance;
     }
 
-    public void setAccount_id(Long account_id) {
+    public void setAccount_id(int account_id) {
         this.account_id = account_id;
     }
 
-    public void setUser_id(Long user_id) {
+    public void setUser_id(int user_id) {
         this.user_id = user_id;
     }
